requirements.txt

numpy==1.22.4
scipy==1.7.3
scikit-learn==1.0.2
opencv-python-headless==4.10.0.84
tqdm==4.67.1
rich==13.8.1
pyyaml==6.0.1
pyquaternion==0.9.9
transforms3d==0.4.2
trimesh==4.4.1
matplotlib==3.5.2
seaborn==0.12.2
tensorboard==2.11.2


chumpy==0.70

fvcore==0.1.5.post20221221
iopath==0.1.10
yacs==0.1.8
omegaconf==2.3.0
timm==0.4.5
torchmetrics==0.11.4
pycocotools==2.0.7
imageio==2.31.2
shapely==1.8.5
